Update text-style.md
guarantee The "This property **doesn't guarteen** the given font will always be set to the text. If the specified font cannot be found at the device, a typeface fallback will occurr and the default typeface will be load." has two spelling mistake,should be "guarteen" to "guarantee","occurr" to "occur".