Update Doc chapter(spelling mistake) (#1176)
* Update text-style.md

guarantee The "This property **doesn't guarteen** the given font will always be set to the text. If the specified font cannot be found at the device, a typeface fallback will occurr and the default typeface will be load." has two spelling mistake,should be "guarteen" to "guarantee","occurr" to "occur".

* Update special-element.md

The "The element serves as content distribution outlest in a composed component template." has a spelling mistake,should be "outlet".

* Update a.md

The "Its purpose and syntax is very simliar to" spelling mistake,should be "similar".

* Update storage.md

The "e.result indicate wether `setItem` is succeed" spelling mistake,should be "whether".

* Update clipboard.md

The "set the text into clipboard, as the same as copy manully." spelling mistake,should be "manually".