* [android] let weex embedders have a chance to decide whether to * initialize v8.
This is necessary if there are several clients share the same v8
as their Javascript engines, such as UC mobile browser and weex
may share the same v8. In that case, UC takes the chance of v8's
initialization and weex no need to do the initialization again.
If weex has its own v8, then weex needs to tell weexv8 to
initialize v8.