[WEEX-74] Added Jira issue number as commit log title
As Apache Jira provides the function that can automatically connect
commit and Jira issue according to the issue number with specific
format, so added a new way in commit message title check regex to
allow this.

Removed the Merge title because we are not going to allow merge
and start to follow the rebase way.

Also updated error message to point to the correct URL.