[WEEX-101][iOS] specify voice-over navigation order is column or vertical
Consider an app that shows items in vertical columns. Normally, voiceOver would navigate through
these items in horizontal rows. Setting the value of this property to YES on the parent element of the
items in the vertical columns causes VoiceOver to respect the app’s grouping and navigate them.And
the default of this property is NO.

Bug:101