[WEEX-158] [iOS] Fix transition's height may show incorrect when selected
XianyuApp's developper found this bug in a detailed case.In a selection page, we found that when multiple choices will trigger a problem, identified, to determine the reason because the data accuracy does not meet the demand.

After fix  with solution by replacing float to double. it is fixed.

So, we fix this bug