[WEEX-270][iOS]WXListComponent should add reload type of data update
Currently WXListComponent data update only supports insertRows. This can't be satisfied under some drop-down loading scenes. For example, the pulldown load needs to be positioned to the original position. When the tableView is still decaling and inserts and tries to fix the position, the cell will blink. Because when deceiring, the externally modified contentOffset will be re-edited by the internal implementation of tableView, altering contentOffset alternately will cause flicker problems. If you use the tableView reloadData this is not a problem, so you need to support the WXListComponent in the drop-down restore location scene to add the reload attribute.

feat：270